add UserService, Role fix AuthController,Card, CardController,CardService, SecurityConfig,User

+++ src/main/java/com/CardManagmentSystem/service/CardService.java
@@ -1,6 +1,7 @@
 package com.CardManagmentSystem.service;
 
 import com.CardManagmentSystem.model.Card;
+import com.CardManagmentSystem.model.CardStatus;
 import com.CardManagmentSystem.repository.CardRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -38,18 +39,39 @@ public class CardService {
         return cardRepository.save(existingCard);
     }
 
-    public void deleteCard(Long id) {
-        // Логика для удаления карты
-        cardRepository.deleteById(id);
+    public boolean deleteCard(Long id) {
+        if (cardRepository.existsById(id)) {
+            cardRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public void transferFunds(Long fromCardId, Long toCardId, BigDecimal amount) {
-        // Логика для перевода средств между картами
-        Card fromCard = getCardById(fromCardId);
-        Card toCard = getCardById(toCardId);
-        fromCard.setBalance(fromCard.getBalance().subtract(amount));
-        toCard.setBalance(toCard.getBalance().add(amount));
-        cardRepository.save(fromCard);
-        cardRepository.save(toCard);
+    public boolean blockCard(Long id) {
+        Card card = getCardById(id);
+        card.setStatus(CardStatus.BLOCKED);
+        cardRepository.save(card);
+        return true;
     }
+
+    public boolean activateCard(Long id) {
+        Card card = getCardById(id);
+        card.setStatus(CardStatus.ACTIVE);
+        cardRepository.save(card);
+        return true;
+    }
+
+    public boolean transferFunds(Long fromId, Long toId, BigDecimal amount) {
+        Card fromCard = getCardById(fromId);
+        Card toCard = getCardById(toId);
+        if (fromCard.getBalance().compareTo(amount) >= 0) {
+            fromCard.setBalance(fromCard.getBalance().subtract(amount));
+            toCard.setBalance(toCard.getBalance().add(amount));
+            cardRepository.save(fromCard);
+            cardRepository.save(toCard);
+            return true;
+        }
+        return false;
+    }
+
 }
